refactor(ui): remove Batch Analysis page; modularize UI into src/ui (header, sidebar, pages, router, state)
