+++ src/ui/__init__.py
@@ -1,0 +1,3 @@
+"""
+UI Component Library for Avalanche eDNA
+"""
