+++ src/ui/components/__init__.py
@@ -1,0 +1,3 @@
+"""
+UI Components Package
+"""
