+++ src/ui/pages/__init__.py
@@ -1,0 +1,3 @@
+"""
+UI Pages Package
+"""
