+++ src/ui/workers/__init__.py
@@ -1,0 +1,3 @@
+"""
+UI Workers Package
+"""
